# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Saturday's Spinout API",
    "description": "REST API for Saturday's Spinout — an iRacing race log and analytics platform.",
    "version": "1.0.0",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/jonsabados/saturdaysspinout"
    }
  },
  "servers": [
    {
      "url": "https://api.saturdaysspinout.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/driver/{driver_id}/analytics": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get race analytics",
        "description": "Aggregated race statistics with optional grouping by dimension or time granularity. `groupBy` and `granularity` are mutually exclusive.",
        "operationId": "getAnalytics",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/DriverID" },
          { "$ref": "#/components/parameters/StartTime" },
          { "$ref": "#/components/parameters/EndTime" },
          {
            "name": "groupBy",
            "in": "query",
            "description": "Dimension to group by (repeatable). Mutually exclusive with granularity.",
            "schema": { "type": "array", "items": { "type": "string", "enum": ["series", "car", "track"] } }
          },
          {
            "name": "granularity",
            "in": "query",
            "description": "Time granularity for time-series breakdown. Mutually exclusive with groupBy.",
            "schema": { "type": "string", "enum": ["day", "week", "month", "year"] }
          },
          { "$ref": "#/components/parameters/SeriesIDFilter" },
          { "$ref": "#/components/parameters/CarIDFilter" },
          { "$ref": "#/components/parameters/TrackIDFilter" }
        ],
        "responses": {
          "200": {
            "description": "Analytics results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": { "$ref": "#/components/schemas/AnalyticsResponse" },
                    "correlationId": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from the /auth/ir/callback endpoint. Contains encrypted iRacing tokens and user claims."
      }
    },
    "parameters": {
      "DriverID": {
        "name": "driver_id",
        "in": "path",
        "required": true,
        "description": "iRacing customer ID. Must match the authenticated user.",
        "schema": { "type": "integer", "format": "int64" }
      },
...
      "StartTime": {
        "name": "startTime",
        "in": "query",
        "required": true,
        "description": "Start of time range (ISO-8601)",
        "schema": { "type": "string", "format": "date-time" }
      },
      "EndTime": {
        "name": "endTime",
        "in": "query",
        "required": true,
        "description": "End of time range (ISO-8601)",
        "schema": { "type": "string", "format": "date-time" }
      },
...
      "SeriesIDFilter": {
        "name": "seriesId",
        "in": "query",
        "description": "Filter by series ID (repeatable; OR within, AND across dimensions)",
        "schema": { "type": "array", "items": { "type": "integer", "format": "int64" } }
      },
      "CarIDFilter": {
        "name": "carId",
        "in": "query",
        "description": "Filter by car ID (repeatable; OR within, AND across dimensions)",
        "schema": { "type": "array", "items": { "type": "integer", "format": "int64" } }
      },
      "TrackIDFilter": {
        "name": "trackId",
        "in": "query",
        "description": "Filter by track ID (repeatable; OR within, AND across dimensions)",
        "schema": { "type": "array", "items": { "type": "integer", "format": "int64" } }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/RequestErrors" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      }
    },
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "correlationId": { "type": "string" }
        }
      },
      "RequestErrors": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": { "type": "string" }
          },
          "fieldErrors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FieldError" }
          },
          "correlationId": { "type": "string" }
        }
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "field": { "type": "string" },
          "error": { "type": "string" },
          "code": { "type": "string" },
          "params": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          }
        }
      },
...
      "AnalyticsResponse": {
        "type": "object",
        "properties": {
          "summary": { "$ref": "#/components/schemas/AnalyticsSummary" },
          "groupedBy": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AnalyticsGroup" }
          },
          "timeSeries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AnalyticsPeriod" }
          }
        }
      },
      "AnalyticsSummary": {
        "type": "object",
        "properties": {
          "raceCount": { "type": "integer" },
          "iRatingStart": { "type": "integer" },
          "iRatingEnd": { "type": "integer" },
          "iRatingDelta": { "type": "integer" },
          "iRatingGain": { "type": "integer" },
          "iRatingLoss": { "type": "integer" },
          "cpiStart": { "type": "number", "format": "double" },
          "cpiEnd": { "type": "number", "format": "double" },
          "cpiDelta": { "type": "number", "format": "double" },
          "cpiGain": { "type": "number", "format": "double" },
          "cpiLoss": { "type": "number", "format": "double" },
          "podiums": { "type": "integer" },
          "top5Finishes": { "type": "integer" },
          "wins": { "type": "integer" },
          "avgFinishPosition": { "type": "number", "format": "double" },
          "avgStartPosition": { "type": "number", "format": "double" },
          "positionsGained": { "type": "number", "format": "double" },
          "totalIncidents": { "type": "integer" },
          "avgIncidents": { "type": "number", "format": "double" }
        }
      },
      "AnalyticsGroup": {
        "type": "object",
        "properties": {
          "seriesId": { "type": "integer", "format": "int64" },
          "carId": { "type": "integer", "format": "int64" },
          "trackId": { "type": "integer", "format": "int64" },
          "summary": { "$ref": "#/components/schemas/AnalyticsSummary" }
        }
      },
      "AnalyticsPeriod": {
        "type": "object",
        "properties": {
          "period": { "type": "string", "description": "Time period label. Format depends on granularity: 2024-01-15 (day), 2024-W03 (week), 2024-01 (month), 2024 (year)." },
          "summary": { "$ref": "#/components/schemas/AnalyticsSummary" }
        }
      },
...
    }
  }
}